TST/CLN: moved the legacy tests out of tseries/test_timeseries.py to tseries/test_timeseries_legacy.py
